@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 namespace ElegantMedia\OxygenFoundation;
 
 use ElegantMedia\OxygenFoundation\Console\Commands\Developer\MovePublicFolderCommand;
@@ -11,78 +13,78 @@ use ElegantMedia\OxygenFoundation\Core\Pathfinder;
 use ElegantMedia\OxygenFoundation\Macros\RegisterResponseMacros;
 use ElegantMedia\OxygenFoundation\Macros\RegisterSchemaMacros;
 use ElegantMedia\OxygenFoundation\Navigation\Navigator;
-use ElegantMedia\OxygenFoundation\Scout\KeywordSearchEngine;
+use ElegantMedia\OxygenFoundation\Scout\Engines\SecureKeywordEngine;
 use Illuminate\Support\ServiceProvider;
 use Laravel\Scout\EngineManager;
 
 class OxygenFoundationServiceProvider extends ServiceProvider
 {
+    use RegisterResponseMacros;
+    use RegisterSchemaMacros;
 
-	use RegisterSchemaMacros;
-	use RegisterResponseMacros;
+    protected array $configFiles = [
+        'oxygen' => __DIR__ . '/../config/oxygen.php',
+        'features' => __DIR__ . '/../config/features.php',
+        'scout' => __DIR__ . '/../config/scout.php',
+    ];
 
-	public function register()
-	{
-		$this->app->singleton('oxygen', function () {
-			return new OxygenCore();
-		});
+    public function register(): void
+    {
+        $this->app->singleton('oxygen', fn () => new OxygenCore);
+        $this->app->singleton(Pathfinder::class);
 
-		$this->app->singleton(Pathfinder::class);
+        foreach ($this->configFiles as $key => $path) {
+            $this->mergeConfigFrom($path, $key);
+        }
 
-		$this->mergeConfigFrom(__DIR__ . '/../config/oxygen.php', 'oxygen');
-		$this->mergeConfigFrom(__DIR__ . '/../config/features.php', 'features');
-		$this->mergeConfigFrom(__DIR__ . '/../config/scout.php', 'scout');
+        $this->registerResponseMacros();
+        $this->registerSchemaMacros();
+        $this->registerCommands();
 
-		$this->registerResponseMacros();
-		$this->registerSchemaMacros();
-		$this->registerCommands();
+        // Register Navigator Facade
+        $this->app->singleton('elegantmedia.oxygen.navigator', fn () => new Navigator);
+    }
 
-		// Register Navigator Facade
-		$this->app->singleton('elegantmedia.oxygen.navigator', function () {
-			return new Navigator();
-		});
-	}
+    public function boot(): void
+    {
+        $this->registerPublishables();
+        $this->bootScoutSearchEngines();
+    }
 
-	public function boot()
-	{
-		// publish config
-		$this->publishes([
-			__DIR__.'/../config/oxygen.php' 	=> config_path('oxygen.php'),
-			__DIR__.'/../config/features.php' 	=> config_path('features.php'),
-			__DIR__.'/../config/scout.php' 		=> config_path('scout.php'),
-		], 'oxygen-config');
+    protected function registerPublishables(): void
+    {
+        if ($this->app->runningInConsole()) {
+            // Publish config files
+            $configPaths = [];
+            foreach ($this->configFiles as $key => $path) {
+                $configPaths[$path] = config_path($key . '.php');
+            }
+            $this->publishes($configPaths, 'oxygen-config');
 
-		$this->publishes([
-			__DIR__.'/../stubs/app' => app_path(),
-		], 'oxygen-foundation-install');
+            // Publish installation stubs
+            $this->publishes([
+                __DIR__ . '/../stubs/app' => app_path(),
+            ], 'oxygen-foundation-install');
+        }
+    }
 
-		$this->bootScoutSearchEngines();
-		// $this->app[EngineManager::class]->extend('keyword', function ($app) {
-		// 	return new KeywordSearchEngine();
-		// });
-	}
+    protected function bootScoutSearchEngines(): void
+    {
+        $this->app[EngineManager::class]->extend('keyword', fn () => new SecureKeywordEngine);
+    }
 
-	protected function bootScoutSearchEngines(): void
-	{
-		$this->app[EngineManager::class]->extend('keyword', function ($app) {
-		// resolve(EngineManager::class)->extend('keyword', function () {
-			return new KeywordSearchEngine();
-		});
-	}
+    protected function registerCommands(): void
+    {
+        $commands = [
+            SeedCommand::class,
+            OxygenFoundationInstallCommand::class,
+            MovePublicFolderCommand::class,
+        ];
 
+        if (! $this->app->environment('production')) {
+            $commands[] = RefreshDatabaseCommand::class;
+        }
 
-	protected function registerCommands(): void
-	{
-		$this->commands([
-			SeedCommand::class,
-			OxygenFoundationInstallCommand::class,
-			MovePublicFolderCommand::class,
-		]);
-
-		if (!app()->environment('production')) {
-			$this->commands([
-				RefreshDatabaseCommand::class,
-			]);
-		}
-	}
+        $this->commands($commands);
+    }
 }
