@@ -1,106 +1,106 @@
 <?php
 
+declare(strict_types=1);
+
 return [
+    /*
+    |--------------------------------------------------------------------------
+    | Default Search Engine
+    |--------------------------------------------------------------------------
+    |
+    | This option controls the default search connection that gets used while
+    | using Laravel Scout. This connection is used when syncing all models
+    | to the search service. You should adjust this based on your needs.
+    |
+    | Supported: "algolia", "null"
+    |
+    */
 
-	/*
-	|--------------------------------------------------------------------------
-	| Default Search Engine
-	|--------------------------------------------------------------------------
-	|
-	| This option controls the default search connection that gets used while
-	| using Laravel Scout. This connection is used when syncing all models
-	| to the search service. You should adjust this based on your needs.
-	|
-	| Supported: "algolia", "null"
-	|
-	*/
+    'driver' => env('SCOUT_DRIVER', 'keyword'),
 
-	'driver' => env('SCOUT_DRIVER', 'keyword'),
+    /*
+    |--------------------------------------------------------------------------
+    | Index Prefix
+    |--------------------------------------------------------------------------
+    |
+    | Here you may specify a prefix that will be applied to all search index
+    | names used by Scout. This prefix may be useful if you have multiple
+    | "tenants" or applications sharing the same search infrastructure.
+    |
+    */
 
-	/*
-	|--------------------------------------------------------------------------
-	| Index Prefix
-	|--------------------------------------------------------------------------
-	|
-	| Here you may specify a prefix that will be applied to all search index
-	| names used by Scout. This prefix may be useful if you have multiple
-	| "tenants" or applications sharing the same search infrastructure.
-	|
-	*/
+    'prefix' => env('SCOUT_PREFIX', ''),
 
-	'prefix' => env('SCOUT_PREFIX', ''),
+    /*
+    |--------------------------------------------------------------------------
+    | Queue Data Syncing
+    |--------------------------------------------------------------------------
+    |
+    | This option allows you to control if the operations that sync your data
+    | with your search engines are queued. When this is set to "true" then
+    | all automatic data syncing will get queued for better performance.
+    |
+    */
 
-	/*
-	|--------------------------------------------------------------------------
-	| Queue Data Syncing
-	|--------------------------------------------------------------------------
-	|
-	| This option allows you to control if the operations that sync your data
-	| with your search engines are queued. When this is set to "true" then
-	| all automatic data syncing will get queued for better performance.
-	|
-	*/
+    'queue' => env('SCOUT_QUEUE', false),
 
-	'queue' => env('SCOUT_QUEUE', false),
+    /*
+    |--------------------------------------------------------------------------
+    | Chunk Sizes
+    |--------------------------------------------------------------------------
+    |
+    | These options allow you to control the maximum chunk size when you are
+    | mass importing data into the search engine. This allows you to fine
+    | tune each of these chunk sizes based on the power of the servers.
+    |
+    */
 
-	/*
-	|--------------------------------------------------------------------------
-	| Chunk Sizes
-	|--------------------------------------------------------------------------
-	|
-	| These options allow you to control the maximum chunk size when you are
-	| mass importing data into the search engine. This allows you to fine
-	| tune each of these chunk sizes based on the power of the servers.
-	|
-	*/
+    'chunk' => [
+        'searchable' => 500,
+        'unsearchable' => 500,
+    ],
 
-	'chunk' => [
-		'searchable' => 500,
-		'unsearchable' => 500,
-	],
+    /*
+    |--------------------------------------------------------------------------
+    | Soft Deletes
+    |--------------------------------------------------------------------------
+    |
+    | This option allows to control whether to keep soft deleted records in
+    | the search indexes. Maintaining soft deleted records can be useful
+    | if your application still needs to search for the records later.
+    |
+    */
 
-	/*
-	|--------------------------------------------------------------------------
-	| Soft Deletes
-	|--------------------------------------------------------------------------
-	|
-	| This option allows to control whether to keep soft deleted records in
-	| the search indexes. Maintaining soft deleted records can be useful
-	| if your application still needs to search for the records later.
-	|
-	*/
+    'soft_delete' => false,
 
-	'soft_delete' => false,
+    /*
+    |--------------------------------------------------------------------------
+    | Identify User
+    |--------------------------------------------------------------------------
+    |
+    | This option allows you to control whether to notify the search engine
+    | of the user performing the search. This is sometimes useful if the
+    | engine supports any analytics based on this application's users.
+    |
+    | Supported engines: "algolia"
+    |
+    */
 
-	/*
-	|--------------------------------------------------------------------------
-	| Identify User
-	|--------------------------------------------------------------------------
-	|
-	| This option allows you to control whether to notify the search engine
-	| of the user performing the search. This is sometimes useful if the
-	| engine supports any analytics based on this application's users.
-	|
-	| Supported engines: "algolia"
-	|
-	*/
+    'identify' => env('SCOUT_IDENTIFY', false),
 
-	'identify' => env('SCOUT_IDENTIFY', false),
+    /*
+    |--------------------------------------------------------------------------
+    | Algolia Configuration
+    |--------------------------------------------------------------------------
+    |
+    | Here you may configure your Algolia settings. Algolia is a cloud hosted
+    | search engine which works great with Scout out of the box. Just plug
+    | in your application ID and admin API key to get started searching.
+    |
+    */
 
-	/*
-	|--------------------------------------------------------------------------
-	| Algolia Configuration
-	|--------------------------------------------------------------------------
-	|
-	| Here you may configure your Algolia settings. Algolia is a cloud hosted
-	| search engine which works great with Scout out of the box. Just plug
-	| in your application ID and admin API key to get started searching.
-	|
-	*/
-
-	'algolia' => [
-		'id' => env('ALGOLIA_APP_ID', ''),
-		'secret' => env('ALGOLIA_SECRET', ''),
-	],
-
+    'algolia' => [
+        'id' => env('ALGOLIA_APP_ID', ''),
+        'secret' => env('ALGOLIA_SECRET', ''),
+    ],
 ];
