@@ -1,4 +1,7 @@
 <?php
+
+declare(strict_types=1);
+
 namespace ElegantMedia\OxygenFoundation\Console\Commands\Traits;
 
 use ElegantMedia\OxygenFoundation\Core\Pathfinder;
@@ -12,100 +15,93 @@ use Symfony\Component\HttpFoundation\File\Exception\FileException;
 
 trait CopiesProjectStubFiles
 {
+    /**
+     * @throws ClassAlreadyExistsException
+     * @throws FileInvalidException
+     * @throws FileNotFoundException
+     */
+    protected function copyMigrationFile($stubPath): string
+    {
+        // because Laravel 5.7 doesn't auto-load migration classes, manually load them
+        // $migrationsPath = database_path('database');
+        // Loader::includeAllFilesFromDir($migrationsPath);
 
+        if (! file_exists($stubPath)) {
+            throw new FileNotFoundException("Filing {$stubPath} not found.");
+        }
 
-	/**
-	 * @param $stubPath
-	 * @return string
-	 * @throws ClassAlreadyExistsException
-	 * @throws FileInvalidException
-	 * @throws FileNotFoundException
-	 */
-	protected function copyMigrationFile($stubPath): string
-	{
-		// because Laravel 5.7 doesn't auto-load migration classes, manually load them
-		// $migrationsPath = database_path('database');
-		// Loader::includeAllFilesFromDir($migrationsPath);
+        $className = FileEditor::getPHPClassName($stubPath);
 
-		if (!file_exists($stubPath)) {
-			throw new FileNotFoundException("Filing {$stubPath} not found.");
-		}
+        $basename = pathinfo($stubPath, PATHINFO_BASENAME);
 
-		$className = FileEditor::getPHPClassName($stubPath);
+        // We'll try to match a few incoming file name formats
+        //
+        // 0001_00_00_00_create_dummies_table.php
+        // 0001_create_dummies_table.php
+        // 0001_00_create_dummies_table.php
+        // 0001_00_00_000_create_dummies_table.php
+        // 0001_00_00_create_dummies_table.php
+        //
+        // From above examples, we have to caputure `create_dummies_table.php` using regex
 
-		$basename = pathinfo($stubPath, PATHINFO_BASENAME);
+        preg_match('/[\d]{1,4}_?(?:\d{1,4}_)+(.*)/', $basename, $matches);
+        if (! is_countable($matches) || count($matches) < 1) {
+            throw new FileInvalidException("Unable to parse migration filename `{$basename}` at `{$stubPath}`.");
+        }
+        $filename = Timing::microTimestamp() . '_' . $matches[1];
 
-		// We'll try to match a few incoming file name formats
-		//
-		// 0001_00_00_00_create_dummies_table.php
-		// 0001_create_dummies_table.php
-		// 0001_00_create_dummies_table.php
-		// 0001_00_00_000_create_dummies_table.php
-		// 0001_00_00_create_dummies_table.php
-		//
-		// From above examples, we have to caputure `create_dummies_table.php` using regex
+        $destinationDir = app(Pathfinder::class)->dbMigrationsDir();
+        File::ensureDirectoryExists($destinationDir);
+        $destinationPath = $destinationDir . DIRECTORY_SEPARATOR . $filename;
 
-		preg_match('/[\d]{1,4}_?(?:\d{1,4}_)+(.*)/', $basename, $matches);
-		if (!is_countable($matches) || count($matches) < 1) {
-			throw new FileInvalidException("Unable to parse migration filename `{$basename}` at `{$stubPath}`.");
-		}
-		$filename = Timing::microTimestamp() . '_' . $matches[1];
+        $this->copyFile($stubPath, $destinationPath, $className);
 
-		$destinationDir = app(Pathfinder::class)->dbMigrationsDir();
-		File::ensureDirectoryExists($destinationDir);
-		$destinationPath = $destinationDir.DIRECTORY_SEPARATOR.$filename;
+        return $destinationPath;
+    }
 
-		$this->copyFile($stubPath, $destinationPath, $className);
+    /**
+     * @throws ClassAlreadyExistsException
+     * @throws FileNotFoundException
+     */
+    protected function copySeedFile($stubPath): string
+    {
+        if (! file_exists($stubPath)) {
+            throw new FileNotFoundException("File {$stubPath} not found.");
+        }
 
-		return $destinationPath;
-	}
+        // stub /tests/TestPack
+        // path /vendor/
+        // path /migrations/seeders/AutoSeed
 
-	/**
-	 * @param $stubPath
-	 * @return string
-	 * @throws ClassAlreadyExistsException
-	 * @throws FileNotFoundException
-	 */
-	protected function copySeedFile($stubPath): string
-	{
-		if (!file_exists($stubPath)) {
-			throw new FileNotFoundException("File {$stubPath} not found.");
-		}
+        $className = FileEditor::getPHPClassName($stubPath);
 
-		// stub /tests/TestPack
-		// path /vendor/
-		// path /migrations/seeders/AutoSeed
+        $filename = pathinfo($stubPath, PATHINFO_BASENAME);
+        $destinationDir = database_path('seeders');
+        File::ensureDirectoryExists($destinationDir);
+        $destinationPath = $destinationDir . DIRECTORY_SEPARATOR . $filename;
 
-		$className = FileEditor::getPHPClassName($stubPath);
+        $result = $this->copyFile($stubPath, $destinationPath, $className);
 
-		$filename = pathinfo($stubPath, PATHINFO_BASENAME);
-		$destinationDir = database_path('seeders');
-		File::ensureDirectoryExists($destinationDir);
-		$destinationPath = $destinationDir.DIRECTORY_SEPARATOR.$filename;
+        return $destinationPath;
+    }
 
-		$result = $this->copyFile($stubPath, $destinationPath, $className);
+    /**
+     * @param null $className
+     *
+     * @throws ClassAlreadyExistsException
+     */
+    protected function copyFile($source, $destination, $className = null): bool
+    {
+        if ($className && class_exists($className, false)) {
+            // $this->warn("{$className} class already exists. Skipped...");
+            // return false;
+            throw new ClassAlreadyExistsException("{$className} class already exists");
+        }
 
-		return $destinationPath;
-	}
+        if (! File::copy($source, $destination)) {
+            throw new FileException("Unable to copy the file {$destination}");
+        }
 
-	/**
-	 * @param $source
-	 * @param $destination
-	 * @param null $className
-	 * @return bool
-	 * @throws ClassAlreadyExistsException
-	 */
-	protected function copyFile($source, $destination, $className = null): bool
-	{
-		if ($className && class_exists($className, false)) {
-			// $this->warn("{$className} class already exists. Skipped...");
-			// return false;
-			throw new ClassAlreadyExistsException("{$className} class already exists");
-		}
-
-		if (!File::copy($source, $destination)) {
-			throw new FileException("Unable to copy the file {$destination}");
-		}
-		return true;
-	}
+        return true;
+    }
 }
