@@ -1,5 +1,6 @@
 <?php
 
+declare(strict_types=1);
 
 namespace ElegantMedia\OxygenFoundation\Entities;
 
@@ -10,41 +11,36 @@ use Illuminate\Http\Request;
 
 class OxygenRepository extends SimpleBaseRepository
 {
+    /**
+     * Fill model data from a request.
+     *
+     * @param null $id
+     */
+    public function fillModelFromRequest(Request $request, $id = null): Model
+    {
+        if (! $id) {
+            $entity = $this->newModel();
+        } else {
+            $entity = $this->find($id);
+        }
 
-	/**
-	 *
-	 * Fill model data from a request
-	 *
-	 * @param Request $request
-	 * @param null $id
-	 *
-	 * @return Model
-	 */
-	public function fillModelFromRequest(Request $request, $id = null): Model
-	{
-		if (!$id) {
-			$entity = $this->newModel();
-		} else {
-			$entity = $this->find($id);
-		}
+        if (! $entity) {
+            throw new ModelNotFoundException;
+        }
 
-		if (!$entity) {
-			throw new ModelNotFoundException();
-		}
+        $data = $request->all();
+        $entity->fill($data);
 
-		$data = $request->all();
-		$entity->fill($data);
+        if (method_exists($this, 'beforeSavingModel')) {
+            $this->beforeSavingModel($request, $entity);
+        }
 
-		if (method_exists($this, 'beforeSavingModel')) {
-			$this->beforeSavingModel($request, $entity);
-		}
+        $entity->save();
 
-		$entity->save();
+        if (method_exists($this, 'afterSavingModel')) {
+            $this->afterSavingModel($request, $entity);
+        }
 
-		if (method_exists($this, 'afterSavingModel')) {
-			$this->afterSavingModel($request, $entity);
-		}
-
-		return $entity->isDirty() ? $entity->refresh() : $entity;
-	}
+        return $entity->isDirty() ? $entity->refresh() : $entity;
+    }
 }
