@@ -1,9 +1,7 @@
 <?php
 
+declare(strict_types=1);
 
 namespace ElegantMedia\OxygenFoundation\Exceptions;
 
-class TokenGenerationException extends \Exception
-{
-
-}
+class TokenGenerationException extends \Exception {}
