@@ -1,9 +1,7 @@
 <?php
 
+declare(strict_types=1);
 
 namespace ElegantMedia\OxygenFoundation\Exceptions;
 
-class UserNotFoundException extends \Exception
-{
-
-}
+class UserNotFoundException extends \Exception {}
