@@ -1,5 +1,6 @@
 <?php
 
+declare(strict_types=1);
 
 namespace ElegantMedia\OxygenFoundation\Facades;
 
@@ -7,15 +8,14 @@ use Illuminate\Support\Collection;
 
 /**
  * @method static Navigator get()
- * @method static NavBar getNavBar(string $navBarName)
+ * @method static NavBar    getNavBar(string $navBarName)
  * @method static Navigator addItem(NavItem $item, $navBarName = 'default');
  * @method static Collection items($navBarName = 'default');
  */
 class Navigator extends \Illuminate\Support\Facades\Facade
 {
-
-	protected static function getFacadeAccessor()
-	{
-		return 'elegantmedia.oxygen.navigator';
-	}
+    protected static function getFacadeAccessor()
+    {
+        return 'elegantmedia.oxygen.navigator';
+    }
 }
