@@ -1,13 +1,14 @@
 <?php
 
+declare(strict_types=1);
+
 namespace ElegantMedia\OxygenFoundation\Http\Traits\Web;
 
 trait CanCRUD
 {
-
-	use FollowsConventions;
-	use CanBrowse;
-	use CanCreate;
-	use CanEdit;
-	use CanDestroy;
+    use CanBrowse;
+    use CanCreate;
+    use CanDestroy;
+    use CanEdit;
+    use FollowsConventions;
 }
