@@ -1,5 +1,6 @@
 <?php
 
+declare(strict_types=1);
 
 namespace ElegantMedia\OxygenFoundation\Http\Traits\Web;
 
@@ -7,26 +8,20 @@ use Illuminate\Http\RedirectResponse;
 
 trait CanDestroy
 {
+    /**
+     * Handle destroy/DELETE method for the controller.
+     */
+    public function destroy($id): ?RedirectResponse
+    {
+        // for safety, you must enable access with `$this->isDestroyAllowed = true;` at the constructor
+        // you should also check for the user's valid permissions before doing this
 
-	/**
-	 *
-	 * Handle destroy/DELETE method for the controller
-	 *
-	 * @param $id
-	 *
-	 * @return RedirectResponse
-	 */
-	public function destroy($id): ?RedirectResponse
-	{
-		// for safety, you must enable access with `$this->isDestroyAllowed = true;` at the constructor
-		// you should also check for the user's valid permissions before doing this
+        if ($this->isDestroyAllowed()) {
+            $this->repo->delete($id);
 
-		if ($this->isDestroyAllowed()) {
-			$this->repo->delete($id);
+            return redirect()->route($this->getIndexRouteName())->with('success', 'Record deleted.');
+        }
 
-			return redirect()->route($this->getIndexRouteName())->with('success', 'Record deleted.');
-		}
-
-		abort(401, 'You are not authorized to access this URL');
-	}
+        abort(401, 'You are not authorized to access this URL');
+    }
 }
