@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 namespace ElegantMedia\OxygenFoundation\Http\Traits\Web;
 
 use ElegantMedia\OxygenFoundation\Entitities\OxygenRepository;
@@ -11,142 +13,138 @@ use Illuminate\Support\Str;
 
 trait FollowsConventions
 {
+    protected $resourceEntityName;
 
-	protected $resourceEntityName;
+    protected $viewsVendorName;
 
-	protected $viewsVendorName;
+    protected $resourcePrefix;
 
-	protected $resourcePrefix;
+    protected $indexRouteName;
 
-	protected $indexRouteName;
+    protected $isDestroyAllowed = false;
 
-	protected $isDestroyAllowed = false;
+    /**
+     * @var OxygenRepository
+     */
+    protected $repo;
 
-	/**
-	 * @var OxygenRepository
-	 */
-	protected $repo;
+    /**
+     * @return mixed
+     */
+    public function getResourceEntityName()
+    {
+        if (empty($this->resourceEntityName)) {
+            throw new \InvalidArgumentException('`$resourceEntityName` is not set.');
+        }
 
-	/**
-	 * @return mixed
-	 */
-	public function getResourceEntityName()
-	{
-		if (empty($this->resourceEntityName)) {
-			throw new \InvalidArgumentException('`$resourceEntityName` is not set.');
-		}
+        return $this->resourceEntityName;
+    }
 
-		return $this->resourceEntityName;
-	}
+    public function getResourceSingularName(): string
+    {
+        return (string) Str::of($this->getResourceEntityName())->singular()->studly();
+    }
 
-	public function getResourceSingularName(): string
-	{
-		return (string) Str::of($this->getResourceEntityName())->singular()->studly();
-	}
+    public function getResourceSingularTitle(): string
+    {
+        return (string) Str::of($this->getResourceEntityName())->singular()->title();
+    }
 
-	public function getResourceSingularTitle(): string
-	{
-		return (string) Str::of($this->getResourceEntityName())->singular()->title();
-	}
+    public function getResourcePluralName(): string
+    {
+        return Str::pluralStudly($this->getResourceEntityName());
+    }
 
-	public function getResourcePluralName(): string
-	{
-		return Str::pluralStudly($this->getResourceEntityName());
-	}
+    /**
+     * @return mixed
+     */
+    public function getResourcePrefix()
+    {
+        return $this->resourcePrefix;
+    }
 
-	/**
-	 * @return mixed
-	 */
-	public function getResourcePrefix()
-	{
-		return $this->resourcePrefix;
-	}
+    public function getResourceKebabName(): string
+    {
+        return Str::kebab($this->getResourceEntityName());
+    }
 
-	public function getResourceKebabName(): string
-	{
-		return Str::kebab($this->getResourceEntityName());
-	}
+    public function isDestroyAllowed(): bool
+    {
+        return $this->isDestroyAllowed;
+    }
 
+    protected function getModel()
+    {
+        if (! $this->repo) {
+            throw new \InvalidArgumentException('Repository is not defined.');
+        }
 
-	public function isDestroyAllowed(): bool
-	{
-		return $this->isDestroyAllowed;
-	}
+        return $this->repo->getModel();
+    }
 
-	protected function getModel()
-	{
-		if (!$this->repo) {
-			throw new \InvalidArgumentException('Repository is not defined.');
-		}
+    /**
+     * @return mixed
+     */
+    protected function getViewsVendorName()
+    {
+        return $this->viewsVendorName;
+    }
 
-		return $this->repo->getModel();
-	}
+    protected function getVendorPrefixedViewName($suffix): string
+    {
+        $suffix = Arr::implodeIgnoreEmpty('.', [
+            $this->getResourcePrefix(),
+            $suffix,
+        ]);
 
-	/**
-	 * @return mixed
-	 */
-	protected function getViewsVendorName()
-	{
-		return $this->viewsVendorName;
-	}
+        $vendorName = $this->getViewsVendorName();
 
-	protected function getVendorPrefixedViewName($suffix): string
-	{
-		$suffix = Arr::implodeIgnoreEmpty('.', [
-			$this->getResourcePrefix(),
-			$suffix,
-		]);
+        if (! empty($vendorName)) {
+            return $vendorName . '::' . $suffix;
+        }
 
-		$vendorName = $this->getViewsVendorName();
+        return $suffix;
+    }
 
-		if (!empty($vendorName)) {
-			return $vendorName.'::'.$suffix;
-		}
+    protected function getIndexRouteName($suffix = 'index'): string
+    {
+        if ($this->indexRouteName) {
+            return $this->indexRouteName;
+        }
 
-		return $suffix;
-	}
+        $route = Arr::implodeIgnoreEmpty('.', [
+            $this->getResourcePrefix(),
+            $this->getResourceKebabName(),
+            $suffix,
+        ]);
 
-	protected function getIndexRouteName($suffix = 'index'): string
-	{
-		if ($this->indexRouteName) {
-			return $this->indexRouteName;
-		}
+        if (! \Illuminate\Support\Facades\Route::has($route)) {
+            throw new \InvalidArgumentException(
+                "Route `$route` is not defined. Create this route or return a valid route from getIndexRouteName()"
+            );
+        }
 
-		$route = Arr::implodeIgnoreEmpty('.', [
-			$this->getResourcePrefix(),
-			$this->getResourceKebabName(),
-			$suffix,
-		]);
+        $this->indexRouteName = $route;
 
-		if (!\Illuminate\Support\Facades\Route::has($route)) {
-			throw new \InvalidArgumentException(
-				"Route `$route` is not defined. Create this route or return a valid route from getIndexRouteName()"
-			);
-		}
+        return $route;
+    }
 
-		$this->indexRouteName = $route;
+    protected function storeOrUpdateRequest(Request $request, $id = null, $rules = null, $messages = null): Model
+    {
+        // validations
+        if ($rules) {
+            $this->validate($request, $rules, $messages ?? []);
+        }
 
-		return $route;
-	}
+        // save and return model
+        return $this->repo->fillModelFromRequest($request, $id);
+    }
 
-	protected function storeOrUpdateRequest(Request $request, $id = null, $rules = null, $messages = null): Model
-	{
-		// validations
-		if ($rules) {
-			$this->validate($request, $rules, $messages ?? []);
-		}
-
-		// save and return model
-		return $this->repo->fillModelFromRequest($request, $id);
-	}
-
-	/**
-	 * @param bool $withDefaults
-	 *
-	 * @return Filterable
-	 */
-	protected function newSearchFilter($withDefaults = true): Filterable
-	{
-		return $this->repo->newSearchFilter($withDefaults);
-	}
+    /**
+     * @param bool $withDefaults
+     */
+    protected function newSearchFilter($withDefaults = true): Filterable
+    {
+        return $this->repo->newSearchFilter($withDefaults);
+    }
 }
