@@ -1,5 +1,6 @@
 <?php
 
+declare(strict_types=1);
 
 namespace ElegantMedia\OxygenFoundation\Navigation;
 
@@ -7,87 +8,72 @@ use Illuminate\Support\Collection;
 
 class NavBar
 {
+    protected $name = Navigator::DEFAULT_NAME;
 
-	protected $name = Navigator::DEFAULT_NAME;
+    protected $items;
 
-	protected $items;
+    public function __construct($navBarName = Navigator::DEFAULT_NAME)
+    {
+        $this->setName($navBarName);
 
-	public function __construct($navBarName = Navigator::DEFAULT_NAME)
-	{
-		$this->setName($navBarName);
+        $this->items = new Collection;
+    }
 
-		$this->items = new Collection();
-	}
+    /**
+     * Add a new NavItem.
+     */
+    public function add(NavItem $item): NavBar
+    {
+        $this->items->push($item);
 
-	/**
-	 *
-	 * Add a new NavItem
-	 *
-	 * @param NavItem $item
-	 *
-	 * @return NavBar
-	 */
-	public function add(NavItem $item): NavBar
-	{
-		$this->items->push($item);
+        return $this;
+    }
 
-		return $this;
-	}
+    /**
+     * Get a list of menu items.
+     *
+     * @return mixed
+     */
+    public function items()
+    {
+        $sorted = $this->items->sort(function ($first, $second) {
+            // ensure orders are numbers, just for safety
+            if (! is_int($first->getOrder())) {
+                $first->setOrder(0);
+            }
 
-	/**
-	 *
-	 * Get a list of menu items
-	 *
-	 * @return mixed
-	 */
-	public function items()
-	{
-		$sorted = $this->items->sort(function ($first, $second) {
+            if (! is_int($second->getOrder())) {
+                $second->setOrder(0);
+            }
 
-			// ensure orders are numbers, just for safety
-			if (!is_int($first->getOrder())) {
-				$first->setOrder(0);
-			}
+            // if the sort order is equal, then sort by text
+            if ($first->getOrder() === $second->getOrder()) {
+                return strcmp(strtolower($first->getText()), strtolower($second->getText()));
+            }
 
-			if (!is_int($second->getOrder())) {
-				$second->setOrder(0);
-			}
+            // otherwise sort by order
+            return (int) ($first->getOrder() > $second->getOrder());
+        });
 
-			// if the sort order is equal, then sort by text
-			if ($first->getOrder() === $second->getOrder()) {
-				return strcmp(strtolower($first->getText()), strtolower($second->getText()));
-			}
+        return $sorted;
+    }
 
-			// otherwise sort by order
-			return (int)($first->getOrder() > $second->getOrder());
-		});
+    public function getItem($itemId)
+    {
+        return $this->items->first(function ($item) use ($itemId) {
+            return $item->getId() === $itemId;
+        });
+    }
 
-		return $sorted;
-	}
+    public function setName(string $name): NavBar
+    {
+        $this->name = $name;
 
-	public function getItem($itemId)
-	{
-		return $this->items->first(function ($item) use ($itemId) {
-			return $item->getId() === $itemId;
-		});
-	}
+        return $this;
+    }
 
-	/**
-	 * @param string $name
-	 *
-	 * @return NavBar
-	 */
-	public function setName(string $name): NavBar
-	{
-		$this->name = $name;
-		return $this;
-	}
-
-	/**
-	 * @return string
-	 */
-	public function getName(): string
-	{
-		return $this->name;
-	}
+    public function getName(): string
+    {
+        return $this->name;
+    }
 }
