@@ -1,8 +1,10 @@
 <?php
 
+declare(strict_types=1);
+
 namespace ElegantMedia\OxygenFoundation\Scout;
 
 interface KeywordSearchable
 {
-	public function getSearchableFields(): array;
+    public function getSearchableFields(): array;
 }
