@@ -1,8 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 namespace ElegantMedia\OxygenFoundation\Support\Exceptions;
 
-class ClassAlreadyExistsException extends \Exception
-{
-
-}
+class ClassAlreadyExistsException extends \Exception {}
