@@ -1,9 +1,7 @@
 <?php
 
+declare(strict_types=1);
 
 namespace ElegantMedia\OxygenFoundation\Support\Exceptions;
 
-class FileInvalidException extends \Exception
-{
-
-}
+class FileInvalidException extends \Exception {}
