@@ -1,4 +1,7 @@
 <?php
+
+declare(strict_types=1);
+
 namespace ElegantMedia\OxygenFoundation\Support;
 
 use Illuminate\Support\Facades\File;
@@ -6,33 +9,32 @@ use Symfony\Component\Finder\SplFileInfo;
 
 class Filing
 {
+    /*
+    |--------------------------------------------------------------------------
+    | Illuminate/Support Functions
+    |--------------------------------------------------------------------------
+    |
+    |
+    |
+    */
 
-	/*
-	|--------------------------------------------------------------------------
-	| Illuminate/Support Functions
-	|--------------------------------------------------------------------------
-	|
-	|
-	|
-	*/
+    public static function allFileNames($directory)
+    {
+        $files = File::allFiles($directory);
 
-	public static function allFileNames($directory)
-	{
-		$files = File::allFiles($directory);
+        return array_map(function ($file) {
+            /* @var SplFileInfo $file */
+            return $file->getPathname();
+        }, $files);
+    }
 
-		return array_map(function ($file) {
-			/* @var SplFileInfo $file */
-			return $file->getPathname();
-		}, $files);
-	}
+    public static function fileNames($directory)
+    {
+        $files = File::files($directory);
 
-	public static function fileNames($directory)
-	{
-		$files = File::files($directory);
-
-		return array_map(function ($file) {
-			/* @var SplFileInfo $file */
-			return $file->getPathname();
-		}, $files);
-	}
+        return array_map(function ($file) {
+            /* @var SplFileInfo $file */
+            return $file->getPathname();
+        }, $files);
+    }
 }
