@@ -1,10 +1,9 @@
 <?php
 
+declare(strict_types=1);
+
 namespace App\Entities;
 
 use ElegantMedia\OxygenFoundation\Entities\OxygenRepository;
 
-class BaseRepository extends OxygenRepository
-{
-
-}
+class BaseRepository extends OxygenRepository {}
