@@ -1,12 +1,12 @@
 <?php
 
+declare(strict_types=1);
+
 namespace App\Http\Controllers\Manage;
 
 use ElegantMedia\OxygenFoundation\Http\Traits\Web\FollowsConventions;
 
 class ManageBaseController extends App\Http\Controllers\Controller
 {
-
-	use FollowsConventions;
-
+    use FollowsConventions;
 }
