@@ -1,33 +1,31 @@
 <?php
 
+declare(strict_types=1);
+
 namespace ElegantMedia\OxygenFoundation\Components\Navigator;
 
 use ElegantMedia\OxygenFoundation\Navigation\NavItem;
 
 class NavItemTest extends \PHPUnit\Framework\TestCase
 {
+    public function test_nav_item_to_array_returns_an_array(): void
+    {
+        $text = 'foo';
 
-	/**
-	 *
-	 */
-	public function testNavItemToArrayReturnsAnArray(): void
-	{
-		$text = "foo";
+        $item = new NavItem;
+        $item->setText($text);
 
-		$item = new NavItem();
-		$item->setText($text);
+        $arr = $item->toArray();
 
-		$arr = $item->toArray();
+        $this->assertEquals($arr['text'], $text);
+    }
 
-		$this->assertEquals($arr["text"], $text);
-	}
+    public function test_nav_item_can_initiate()
+    {
+        $item = new NavItem('My Menu');
 
-	public function testNavItemCanInitiate()
-	{
-		$item = new NavItem('My Menu');
+        $array = $item->toArray();
 
-		$array = $item->toArray();
-
-		$this->assertEquals($array['text'], 'My Menu');
-	}
+        $this->assertEquals($array['text'], 'My Menu');
+    }
 }
