@@ -1,5 +1,6 @@
 <?php
 
+declare(strict_types=1);
 
 namespace ElegantMedia\OxygenFoundation\Tests\Feature;
 
@@ -10,28 +11,27 @@ use ElegantMedia\OxygenFoundation\TestPackage\TestPackageServiceProvider;
 
 class TestCase extends \Orchestra\Testbench\TestCase
 {
+    protected $pathfinder;
 
-	protected $pathfinder;
+    public function __construct(string $name)
+    {
+        parent::__construct($name);
 
-	public function __construct(string $name)
-	{
-		parent::__construct($name);
+        $this->pathfinder = app(Pathfinder::class);
+    }
 
-		$this->pathfinder = app(Pathfinder::class);
-	}
+    protected function getPackageProviders($app)
+    {
+        return [
+            OxygenFoundationServiceProvider::class,
+            TestPackageServiceProvider::class,
+        ];
+    }
 
-	protected function getPackageProviders($app)
-	{
-		return [
-			OxygenFoundationServiceProvider::class,
-			TestPackageServiceProvider::class
-		];
-	}
-
-	protected function getPackageAliases($app)
-	{
-		return [
-			'Navigator' => Navigator::class,
-		];
-	}
+    protected function getPackageAliases($app)
+    {
+        return [
+            'Navigator' => Navigator::class,
+        ];
+    }
 }
