@@ -1,33 +1,34 @@
 <?php
 
+declare(strict_types=1);
+
 use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 
 class CreateDummiesTable extends Migration
 {
-	/**
-	 * Run the database.
-	 *
-	 * @return void
-	 */
-	public function up()
-	{
-		Schema::create('dummies', static function (Blueprint $table) {
-			$table->id('id');
-			$table->string('name')->nullable();
-			$table->softDeletes();
-			$table->timeStamps();
-		});
-	}
+    /**
+     * Run the database.
+     *
+     * @return void
+     */
+    public function up()
+    {
+        Schema::create('dummies', static function (Blueprint $table) {
+            $table->id('id');
+            $table->string('name')->nullable();
+            $table->softDeletes();
+            $table->timeStamps();
+        });
+    }
 
-
-	/**
-	 * Reverse the database.
-	 *
-	 * @return void
-	 */
-	public function down()
-	{
-		Schema::drop('dummies');
-	}
+    /**
+     * Reverse the database.
+     *
+     * @return void
+     */
+    public function down()
+    {
+        Schema::drop('dummies');
+    }
 }
