@@ -1,10 +1,9 @@
 <?php
 
+declare(strict_types=1);
+
 namespace App\Entities\Dummies;
 
 use Illuminate\Database\Eloquent\Model;
 
-class Dummy extends Model
-{
-
-}
+class Dummy extends Model {}
