@@ -1,10 +1,9 @@
 <?php
 
+declare(strict_types=1);
+
 namespace Database\Seeders\OxygenExtensions\AutoSeed;
 
 use Illuminate\Database\Seeder;
 
-class TestAutoSeeder extends Seeder
-{
-
-}
+class TestAutoSeeder extends Seeder {}
