@@ -1,18 +1,20 @@
 <?php
 
+declare(strict_types=1);
+
 namespace Database\Seeders;
 
 use Illuminate\Database\Seeder;
 
 class TestSeeder extends Seeder
 {
-	/**
-	 * Run the database seeders.
-	 *
-	 * @return void
-	 */
-	public function run()
-	{
-		//
-	}
+    /**
+     * Run the database seeders.
+     *
+     * @return void
+     */
+    public function run()
+    {
+        //
+    }
 }
