@@ -1,4 +1,7 @@
 <?php
+
+declare(strict_types=1);
+
 namespace ElegantMedia\OxygenFoundation\TestPackage\Console;
 
 use ElegantMedia\OxygenFoundation\Console\Commands\ExtensionInstallCommand;
@@ -6,17 +9,17 @@ use ElegantMedia\OxygenFoundation\TestPackage\TestPackageServiceProvider;
 
 class TestExtensionInstallCommand extends ExtensionInstallCommand
 {
+    protected $signature = 'setup:extension:test-extension';
 
-	protected $signature 	= 'setup:extension:test-extension';
-	protected $description 	= 'Oxygen Test Extension';
+    protected $description = 'Oxygen Test Extension';
 
-	public function getExtensionServiceProvider(): string
-	{
-		return TestPackageServiceProvider::class;
-	}
+    public function getExtensionServiceProvider(): string
+    {
+        return TestPackageServiceProvider::class;
+    }
 
-	public function getExtensionDisplayName(): string
-	{
-		return 'TestExtension';
-	}
+    public function getExtensionDisplayName(): string
+    {
+        return 'TestExtension';
+    }
 }
