@@ -1,9 +1,7 @@
 <?php
 
+declare(strict_types=1);
 
 namespace ElegantMedia\OxygenFoundation\TestPackage\Entities\Testers;
 
-class Tester extends \Illuminate\Database\Eloquent\Model
-{
-
-}
+class Tester extends \Illuminate\Database\Eloquent\Model {}
