@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 namespace Tests\Traits;
 
 use ElegantMedia\OxygenFoundation\Scout\KeywordSearchEngine;
@@ -9,19 +11,18 @@ use Mockery\MockInterface;
 
 trait MocksScoutEngines
 {
+    use InteractsWithContainer;
 
-	use InteractsWithContainer;
+    protected function mockScoutKeywordEngine(): MockInterface
+    {
+        $engineMock = mock(app(KeywordSearchEngine::class))->makePartial()->shouldIgnoreMissing();
 
-	protected function mockScoutKeywordEngine(): MockInterface
-	{
-		$engineMock = mock(app(KeywordSearchEngine::class))->makePartial()->shouldIgnoreMissing();
+        $managerMock = mock(EngineManager::class)->makePartial()->shouldIgnoreMissing();
 
-		$managerMock = mock(EngineManager::class)->makePartial()->shouldIgnoreMissing();
+        $managerMock->shouldReceive('driver')->andReturn($engineMock);
 
-		$managerMock->shouldReceive('driver')->andReturn($engineMock);
+        $this->swap(EngineManager::class, $managerMock);
 
-		$this->swap(EngineManager::class, $managerMock);
-
-		return $engineMock;
-	}
+        return $engineMock;
+    }
 }
