@@ -1,10 +1,9 @@
 <?php
 
+declare(strict_types=1);
+
 namespace App\Entities\Examples;
 
 use Illuminate\Database\Eloquent\Model;
 
-class Example extends Model
-{
-
-}
+class Example extends Model {}
